login functionality completed
