+++ routers/userRouter.mjs
@@ -10,5 +10,6 @@ const userCtrl = new UserCtrl('main', db.User, db);
 // router.post('/logout', userCtrl.getMain.bind(userCtrl));
 router.post('/register', userCtrl.postRegister.bind(userCtrl));
 router.post('/login', userCtrl.postLogin.bind(userCtrl));
+router.post('/email', userCtrl.postEmail.bind(userCtrl));
 
 export default router;
